Dining 완성

+++ backend/src/main/java/com/finalproject/possystem/table/DiningService.java
@@ -5,6 +5,7 @@ import com.finalproject.possystem.table.repository.DiningRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.Collections;
 import java.util.List;
 
 @Service
@@ -28,13 +29,36 @@ public class DiningService {
     }
 
     /* 테이블을 생성하는 메서드 */
-    public List<Dining> newTable(List<Dining> dining){
+    public Dining saveTable(Dining dining){
+        return diningRepo.save(dining);
+    }
+
+    public List<Dining> saveTableAll(List<Dining> dining) {
         return diningRepo.saveAll(dining);
     }
 
     /* 하나의 테이블을 삭제 */
     public void delTable(Integer tableNo){
-        diningRepo.deleteById(tableNo);
+        if (tableNo != null) {
+            diningRepo.deleteById(tableNo);
+        } else {
+            throw new IllegalArgumentException("Table number must not be null");
+        }
+    }
+
+    /* DB에 비어있는 테이블 찾기 */
+    public Integer findNextTableNo(){
+        List<Integer> tableNos = diningRepo.findAllTableNo();
+        Collections.sort(tableNos); // 테이블번호 정렬
+
+        int nextno = 1; /* 1번부터 시작 */
+        for(Integer tableNo : tableNos) {
+            if(tableNo != nextno){
+                break; /* 비어있는 번호를 찾으면 종료 */
+            }
+            nextno++;
+        }
+        return nextno;
     }
 
     /* table 수정할경우 테이블색상과 위치 변경하는 update문 실행 */
